3/29 업데이트

+++ hello-spring/src/main/java/com/hello/forum/bbs/web/BoardController.java
@@ -1,22 +1,38 @@
 package com.hello.forum.bbs.web;
 
+
+
+import java.io.File;
+import java.io.FileOutputStream;
+import java.io.IOException;
+import java.io.OutputStream;
+import java.util.List;
+
+import org.apache.poi.ss.usermodel.Cell;
+import org.apache.poi.ss.usermodel.Row;
+import org.apache.poi.ss.usermodel.Sheet;
+import org.apache.poi.ss.usermodel.Workbook;
+import org.apache.poi.xssf.streaming.SXSSFWorkbook;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.io.Resource;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.SessionAttribute;
 import org.springframework.web.multipart.MultipartFile;
 
 import com.hello.forum.bbs.service.BoardService;
 import com.hello.forum.bbs.vo.BoardListVO;
 import com.hello.forum.bbs.vo.BoardVO;
 import com.hello.forum.beans.FileHandler;
+import com.hello.forum.member.vo.MemberVO;
 import com.hello.forum.utils.ValidationUtils;
+
+import jakarta.servlet.http.HttpSession;
 
 //import jakarta.validation.Valid;
 
@@ -51,7 +67,10 @@ public class BoardController {
 	 * @return
 	 */
 	@GetMapping("/board/write") // 브라우저에서 링크를 클릭, 브라우저 URL을 직접 입력.
-	public String viewBoardWritePage() {
+	public String viewBoardWritePage(HttpSession session) {
+		
+		
+		
 		return "board/boardwrite";
 	}
 
@@ -91,7 +110,8 @@ public class BoardController {
 			 * @Valid 에 의해 실행된 파라미터 검사 (@Not Empty, Email, Min 등 )의 결과. BindingResult에 저장된다.
 			 */
 //			BindingResult bindingResult,
-			@RequestParam MultipartFile file, Model model) {
+			@RequestParam MultipartFile file, Model model,
+			@SessionAttribute("_LOGIN_USER_") MemberVO memberVO) {
 		System.out.println("글 등록 처리를 해야합니다.");
 		/*
 		 * Servlet Like HttpServletRequest를 이용. - Interceptor에서 이용 - Filter에서 이용.
@@ -115,7 +135,7 @@ public class BoardController {
 		// 수동 검사 시작.
 		// 제목 검사.
 		boolean isNotEmptySubject = ValidationUtils.notEmpty(boardVO.getSubject());
-		boolean isNotEmptyEmail = ValidationUtils.notEmpty(boardVO.getEmail());
+		
 		boolean isNotEmptyContent = ValidationUtils.notEmpty(boardVO.getContent());
 		boolean isEmailFormat = ValidationUtils.email(boardVO.getEmail());
 
@@ -126,12 +146,7 @@ public class BoardController {
 			return "board/boardwrite";
 		}
 
-		if (!isNotEmptyEmail) {
-			// 이메일을 입력하지 않았다면.
-			model.addAttribute("errorMessage", "이메일은 필수 입력 값입니다.");
-			model.addAttribute("boardVO", boardVO);
-			return "board/boardwrite";
-		}
+		
 
 		if (!isNotEmptyContent) {
 			// 내용을 입력하지 않았다면.
@@ -140,12 +155,8 @@ public class BoardController {
 			return "board/boardwrite";
 		}
 
-		if (!isEmailFormat) {
-			// 이메일을 이메일 형태로 입력하지 않았다면.
-			model.addAttribute("errorMessage", "이메일을 올바른 형태로 작성해주세요.");
-			model.addAttribute("boardVO", boardVO);
-			return "board/boardwrite";
-		}
+		
+		boardVO.setEmail(memberVO.getEmail());
 
 		boolean isCreateSuccess = this.boardService.createNewBoard(boardVO, file);
 		if (isCreateSuccess) {
@@ -184,9 +195,14 @@ public class BoardController {
 	}
 
 	@GetMapping("/board/modify/{id}") // /board/modify/1 <-- id 변수의 값은 1
-	public String viewBoardModifyPage(@PathVariable int id, Model model) {
+	public String viewBoardModifyPage(@PathVariable int id, Model model,
+			@SessionAttribute("_LOGIN_USER_") MemberVO memberVO) {
 		// 1. 전달받은 id의 값으로 게시글을 조회한다.
 		BoardVO boardVO = this.boardService.getOneBoard(id, false);
+		
+		if ( ! memberVO.getEmail().equals(boardVO.getEmail())) {
+			throw new IllegalArgumentException("잘못된 접근입니다.");
+		}
 
 		// 2. 게시글의 정보를 화면에 보내준다.
 		model.addAttribute("boardVO", boardVO);
@@ -203,14 +219,21 @@ public class BoardController {
 	 * @return
 	 */
 	@PostMapping("/board/modify/{id}")
-	public String doBoardModify(Model model, @PathVariable int id, BoardVO boardVO, @RequestParam MultipartFile file) {
+	public String doBoardModify(Model model, @PathVariable int id, BoardVO boardVO, @RequestParam MultipartFile file, 
+			@SessionAttribute("_LOGIN_USER_") MemberVO memberVO) {
 
+		BoardVO originalBoardVO = this.boardService.getOneBoard(id, false);
+		
+		if (originalBoardVO.getEmail().equals(memberVO.getEmail())) {
+			throw new IllegalArgumentException("잘못된 접근입니다.");
+		}
+		
 		// 수동 검사 시작.
 		// 제목 검사.
 		boolean isNotEmptySubject = ValidationUtils.notEmpty(boardVO.getSubject());
-		boolean isNotEmptyEmail = ValidationUtils.notEmpty(boardVO.getEmail());
+//		boolean isNotEmptyEmail = ValidationUtils.notEmpty(boardVO.getEmail());
 		boolean isNotEmptyContent = ValidationUtils.notEmpty(boardVO.getContent());
-		boolean isEmailFormat = ValidationUtils.email(boardVO.getEmail());
+//		boolean isEmailFormat = ValidationUtils.email(boardVO.getEmail());
 
 		if (!isNotEmptySubject) {
 			// 제목을 입력하지 않았다면.
@@ -219,12 +242,12 @@ public class BoardController {
 			return "board/boardmodify";
 		}
 
-		if (!isNotEmptyEmail) {
-			// 이메일을 입력하지 않았다면.
-			model.addAttribute("errorMessage", "이메일은 필수 입력 값입니다.");
-			model.addAttribute("boardVO", boardVO);
-			return "board/boardmodify";
-		}
+//		if (!isNotEmptyEmail) {
+//			// 이메일을 입력하지 않았다면.
+//			model.addAttribute("errorMessage", "이메일은 필수 입력 값입니다.");
+//			model.addAttribute("boardVO", boardVO);
+//			return "board/boardmodify";
+//		}
 
 		if (!isNotEmptyContent) {
 			// 내용을 입력하지 않았다면.
@@ -233,16 +256,17 @@ public class BoardController {
 			return "board/boardmodify";
 		}
 
-		if (!isEmailFormat) {
-			// 이메일을 이메일 형태로 입력하지 않았다면.
-			model.addAttribute("errorMessage", "이메일을 올바른 형태로 작성해주세요.");
-			model.addAttribute("boardVO", boardVO);
-			return "board/boardmodify";
-		}
+//		if (!isEmailFormat) {
+//			// 이메일을 이메일 형태로 입력하지 않았다면.
+//			model.addAttribute("errorMessage", "이메일을 올바른 형태로 작성해주세요.");
+//			model.addAttribute("boardVO", boardVO);
+//			return "board/boardmodify";
+//		}
 
 		// Command Object 에는 전달된 ID가 없으므로
 		// PathVariable로 전달된 ID를 셋팅해준다.
 		boardVO.setId(id);
+//		boardVO.setEmail(memberVO.getEmail());
 
 		boolean isUpdatedSuccess = this.boardService.updateOneBoard(boardVO, file);
 
@@ -267,8 +291,14 @@ public class BoardController {
 	 * FORM 으로 데이터를 등록하거나 수정할 경우 -> POST URL이나 링크 등으로 데이터를 조회하거나 삭제할 경우 -> GET
 	 */
 	@GetMapping("/board/delete/{id}")
-	public String doDeleteBoard(@PathVariable int id) {
+	public String doDeleteBoard(@PathVariable int id, @SessionAttribute("_LOGIN_USER_") MemberVO memberVO) {
 
+		BoardVO OriginalBoardVO = this.boardService.getOneBoard(id, false);
+		
+		if ( !OriginalBoardVO.getEmail().equals(memberVO.getEmail())) {
+			throw new IllegalArgumentException("잘못된 접근입니다.");
+		}
+		
 		boolean isDeletedSuccess = this.boardService.deleteOneBoard(id);
 
 		if (isDeletedSuccess) {
@@ -295,7 +325,105 @@ public class BoardController {
 		}
 		// 첨부된 파일이 있을 경우엔 파일을 사용자에게 보내준다. (Downlaod)
 		// 사용자가 업로드한 파일 원본이름 // DB에 저장된 난수화된 파일이름
-		return this.fileHandler.downlaod(boardVO.getOriginFileName(), boardVO.getFileName());
+		return this.fileHandler.download(boardVO.getOriginFileName(), boardVO.getFileName());
+	}
+	
+	@GetMapping("/board/excel/download")
+	public ResponseEntity<Resource> downloadExcelFile() {
+		
+		// 모든 게시글 조회
+		BoardListVO boardListVO = boardService.getAllBoard();
+		
+		// XLSX 문서 만들기
+		Workbook workbook = new SXSSFWorkbook(-1);
+		
+		// 엑셀 시트 만들기
+		Sheet sheet = workbook.createSheet("게시글 목록");
+		
+		// 행 만들기
+		Row row = sheet.createRow(0);
+		
+		// 타이틀 만들기
+		Cell cell = row.createCell(0);
+		cell.setCellValue("번호");
+		
+		cell = row.createCell(1);
+		cell.setCellValue("제목");
+		
+		cell = row.createCell(2);
+		cell.setCellValue("첨부파일명");
+		
+		cell = row.createCell(3);
+		cell.setCellValue("작성자이메일");
+		
+		cell = row.createCell(4);
+		cell.setCellValue("조회수");
+		
+		cell = row.createCell(5);
+		cell.setCellValue("등록일");
+		
+		cell = row.createCell(6);
+		cell.setCellValue("수정일");
+		
+		// 데이터 행 만들고 쓰기
+		List<BoardVO> boardList = boardListVO.getBoardList();
+		int rowIndex = 1;
+		
+		for (BoardVO boardVO : boardList) {
+			row = sheet.createRow(rowIndex);
+			cell = row.createCell(0);
+			cell.setCellValue("" + boardVO.getId());
+			
+			cell = row.createCell(1);
+			cell.setCellValue(boardVO.getSubject());
+			
+			cell = row.createCell(2);
+			cell.setCellValue(boardVO.getOriginFileName());
+			
+			cell = row.createCell(3);
+			cell.setCellValue(boardVO.getEmail());
+		
+			cell = row.createCell(4);
+			cell.setCellValue(boardVO.getViewCnt());
+			
+			cell = row.createCell(5);
+			cell.setCellValue(boardVO.getCrtDt());
+			
+			cell = row.createCell(6);
+			cell.setCellValue(boardVO.getMdfyDt());
+			
+			rowIndex += 1;
+			
+		}
+		
+		// 엑셀 파일 만들기
+		File storedFile = fileHandler.getStroedFile("게시글_목록.xlsx");
+		OutputStream os = null;
+		
+		try {
+			os = new FileOutputStream(storedFile);
+			workbook.write(os);
+		} catch (IOException e) {
+			throw new IllegalArgumentException("엑셀 파일을 만들 수 없습니다.");
+		} finally {
+			try {
+				workbook.close();
+			} catch (IOException e) {} 
+			if (os != null) {
+				try {
+					os.flush();
+				} catch (IOException e) {}
+				try {
+					os.close();
+				} catch (IOException e) {}
+			}
+		}
+		return this.fileHandler.download("게시글_목록.xlsx", "게시글_목록.xlsx");
 	}
 
 }
+
+
+
+
+
